Fixed: Typo in scripting interface. MaxSlotsLimt -> MaxSlotsLimit (thx Cęńoßy†ę for report).
git-svn-id: svn://svn.czdc.org/PtokaX@155 4c629181-62e4-40b4-9e0d-b9f719119ca2

+++ core/DB-SQLite.cpp
@@ -232,7 +232,7 @@ static int SelectCallBack(void *, int iArgCount, char ** ppArgSTrings, char **) 
 
 		szLength = strlen(ppArgSTrings[0]);
 		if(szLength == 0 || szLength > 64) {
-			UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid nick length: %" PRIu64, (uint64_t)szLength);
+			UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid nick length: %zu", szLength);
 			return 0;
 		}
 
@@ -352,13 +352,13 @@ static int SelectCallBack(void *, int iArgCount, char ** ppArgSTrings, char **) 
 
 			szLength = strlen(ppArgSTrings[2]);
 			if(szLength == 0 || szLength > 39) {
-				UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid ip length: %" PRIu64, (uint64_t)szLength);
+				UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid ip length: %zu", szLength);
 				return 0;
 			}
 
 			szLength = strlen(ppArgSTrings[3]);
 			if(szLength == 0 || szLength > 24) {
-				UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid share length: %" PRIu64, (uint64_t)szLength);
+				UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid share length: %zu", szLength);
 				return 0;
 			}
 
@@ -373,7 +373,7 @@ static int SelectCallBack(void *, int iArgCount, char ** ppArgSTrings, char **) 
 			szLength = strlen(ppArgSTrings[4]);
             if(szLength != 0) {
 				if(szLength > 192) {
-					UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid description length: %" PRIu64, (uint64_t)szLength);
+					UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid description length: %zu", szLength);
 					return 0;
 				}
 
@@ -389,7 +389,7 @@ static int SelectCallBack(void *, int iArgCount, char ** ppArgSTrings, char **) 
 			szLength = strlen(ppArgSTrings[5]);
             if(szLength != 0) {
 				if(szLength > 192) {
-					UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid tag length: %" PRIu64, (uint64_t)szLength);
+					UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid tag length: %zu", szLength);
 					return 0;
 				}
 
@@ -405,7 +405,7 @@ static int SelectCallBack(void *, int iArgCount, char ** ppArgSTrings, char **) 
 			szLength = strlen(ppArgSTrings[6]);
             if(szLength != 0) {
 				if(szLength > 32) {
-					UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid connection length: %" PRIu64, (uint64_t)szLength);
+					UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid connection length: %zu", szLength);
 					return 0;
 				}
 
@@ -421,7 +421,7 @@ static int SelectCallBack(void *, int iArgCount, char ** ppArgSTrings, char **) 
 			szLength = strlen(ppArgSTrings[7]);
             if(szLength != 0) {
 				if(szLength > 96) {
-					UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBPostgreSQL search returned invalid email length: %" PRIu64, (uint64_t)szLength);
+					UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBPostgreSQL search returned invalid email length: %zu", szLength);
 					return 0;
 				}
 
@@ -455,13 +455,13 @@ static int SelectCallBack(void *, int iArgCount, char ** ppArgSTrings, char **) 
 
 		size_t szLength = strlen(sFirstNick);
 		if(szLength == 0 || szLength > 64) {
-			UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid nick length: %" PRIu64, (uint64_t)szLength);
+			UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid nick length: %zu", szLength);
 			return 0;
 		}
 	
 		szLength = strlen(sFirstIP);
 		if(szLength == 0 || szLength > 39) {
-			UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid ip length: %" PRIu64, (uint64_t)szLength);
+			UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid ip length: %zu", szLength);
 			return 0;
 		}
 	
@@ -474,13 +474,13 @@ static int SelectCallBack(void *, int iArgCount, char ** ppArgSTrings, char **) 
 
 	size_t szLength = strlen(ppArgSTrings[0]);
 	if(szLength == 0 || szLength > 64) {
-		UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid nick length: %" PRIu64, (uint64_t)szLength);
+		UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid nick length: %zu", szLength);
 		return 0;
 	}
 
 	szLength = strlen(ppArgSTrings[2]);
 	if(szLength == 0 || szLength > 39) {
-		UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid ip length: %" PRIu64, (uint64_t)szLength);
+		UdpDebug::m_Ptr->BroadcastFormat("[LOG] DBSQLite search returned invalid ip length: %zu", szLength);
 		return 0;
 	}
