+++ core/ProfileManager.cpp
@@ -576,7 +576,7 @@ ProfileItem * ProfileManager::CreateProfile(const char * sName) {
 	pNewProfile->m_sName = (char *)malloc(szLen+1);
 #endif
     if(pNewProfile->m_sName == NULL) {
-		AppendDebugLogFormat("[MEM] Cannot allocate %" PRIu64 " bytes in ProfileManager::CreateProfile for pNewProfile->sName\n", (uint64_t)szLen);
+		AppendDebugLogFormat("[MEM] Cannot allocate %zu bytes in ProfileManager::CreateProfile for pNewProfile->sName\n", szLen);
 
         exit(EXIT_FAILURE);
     } 
@@ -716,7 +716,7 @@ void ProfileManager::ChangeProfileName(const uint16_t ui16Profile, char * sName,
     if(m_ppProfilesTable[ui16Profile]->m_sName == NULL) {
         m_ppProfilesTable[ui16Profile]->m_sName = sOldName;
 
-		AppendDebugLogFormat("[MEM] Cannot reallocate %" PRIu64 " bytes in ProfileManager::ChangeProfileName for m_ppProfilesTable[ui16Profile]->sName\n", (uint64_t)szLen);
+		AppendDebugLogFormat("[MEM] Cannot reallocate %zu bytes in ProfileManager::ChangeProfileName for m_ppProfilesTable[ui16Profile]->sName\n", szLen);
 
         return;
     }
